build: Move image building to cmd/build.go
Move the image building code from QEMU driver to cmd/build.go. This is
the first step into killing the QEMU dependency for building images and
use whatever hypervisor is available on the system.
